requirements.txt
openai~=0.27.6
python-dotenv==1.0.0
tweepy~=4.14.0
praw~=7.7.0